% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{chain_to_mcmc}
\alias{chain_to_mcmc}
\title{Extract the parameters from a list of moprobit state containers}
\usage{
chain_to_mcmc(trace)
}
\arguments{
\item{trace}{A list of moprobit state containers}
}
\value{
A matrix of parameter values, rows corresponding to iterations and columns to parameters
}
\description{
Extract the parameters from a list of moprobit state containers
}
\seealso{
\code{\link[=moprobit_chain]{moprobit_chain()}}, \code{\link[=moprobit_iter]{moprobit_iter()}}, \code{\link[=get_toc]{get_toc()}}, \code{\link[=as.mcmc]{as.mcmc()}}
}
