% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{clean_state}
\alias{clean_state}
\title{Remove the derived quantities from state}
\usage{
clean_state(state)
}
\arguments{
\item{state}{A moprobit state container.}
}
\value{
The state container with derived quantities removed.
}
\description{
Remove the derived quantities from state
}
\seealso{
\code{\link[=derive_state]{derive_state()}}
}
