% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{derive_state}
\alias{derive_state}
\title{Recalculate the derived quantities for state}
\usage{
derive_state(state)
}
\arguments{
\item{state}{A moprobit state container.}
}
\value{
The state container with the derived quantities (\code{Y}, \code{X}, \code{mu}, and \code{E}) recomputed.
}
\description{
Recalculate the derived quantities for state
}
\seealso{
\code{\link[=clean_state]{clean_state()}}
}
