% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_acf}
\alias{get_acf}
\title{Get the autocorrelation of the MCMC parameters}
\usage{
get_acf(par)
}
\arguments{
\item{par}{A matrix of parameters, from \code{\link[=chain_to_mcmc]{chain_to_mcmc()}}}
}
\value{
FIXME
}
\description{
Get the autocorrelation of the MCMC parameters
}
