% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_toc}
\alias{get_toc}
\title{Obtain metainformation about the multivariate ordinal probit model parameters}
\usage{
get_toc(par, env = NULL)
}
\arguments{
\item{par}{A matrix returned from \code{\link[=chain_to_mcmc]{chain_to_mcmc()}}}

\item{env}{The MCMC environment from a state container returned from \code{\link[=moprobit_init]{moprobit_init()}}}
}
\value{
A data frame describing the parameters
}
\description{
Obtain metainformation about the multivariate ordinal probit model parameters
}
\details{
FIXME
}
\seealso{
\code{\link[=chain_to_mcmc]{chain_to_mcmc()}}
}
