% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{moprobit_init}
\alias{moprobit_init}
\title{Multivariate ordinal probit MCMC initialization}
\usage{
moprobit_init(formulas, dataset, sd_tau, meas_err = NULL, contrasts = NULL,
  subset = NULL, initZero = FALSE)
}
\arguments{
\item{formulas}{A formula or a named list of formulas specifying blocks of outcomes with common specification. See details.}

\item{dataset}{The data to be used in estimating the model.}

\item{sd_tau}{The standard deviation of the proposal distribution for \code{tau} parameters.}

\item{meas_err}{A named list of measurement error variances for continuous variables measured with error. Each element should have the same name as a variable in \code{formulas} and should either be a scalar or a vector of the same length as \code{dataset}.}

\item{contrasts}{A named list of factor contrasts. See the \code{contrasts.arg} of \code{\link[=model.matrix.default]{model.matrix.default()}}.}

\item{subset}{Either a logical vector of cases in \code{dataset} to use for estimation of the entire model, or a named list of logical vectors indicating which cases should be used to estimate each block defined by \code{formulas}. Note that you must call \code{\link[=moprobit_iter]{moprobit_iter()}} with \code{fixCrossBlockCov = TRUE} if you subset by block.}

\item{initZero}{If \code{TRUE}, initialize \code{beta} with 0 and \code{tau} with whole numbers. Otherwise, initial values will be obtained from univariate models.}
}
\value{
A moprobit state container for use by \code{\link[=moprobit_iter]{moprobit_iter()}}
}
\description{
Initialize a new MCMC chain for a multivariate ordinal probit model
}
\details{
Formulas may contain main effects and interaction terms specified using \code{*} or \code{:}, as with \code{\link[=lm]{lm()}}.
Do not include higher-order terms or computed terms using \code{\link[=I]{I()}}.
Outcome variables, on the left side of the formula's \code{~} should be concatenated with \code{+} (note that this usage
differs from the \code{cbind} convention of \code{\link[=lm]{lm()}}). A variable may be included as an outcome variable (on the left side)
in only one block. A variable may be included as an outcome variable and a predictor variable (on the right side), but
variables should generally be included as predictors only in blocks that come \emph{after} the block in which the variable
is an outcome. Outcome variables may have missing data, but predictors that are not also outcomes may not have any
missing data in \code{dataset}. One strategy for handling missing data in predictors is to include a "base block" that lists
as outcomes any variables with missing data that would otherwise only serve as predictors. This base block may either
include only an intercept on the right (\code{~ 1}) or may include the set of predictors for which there is no missing data.
}
\seealso{
\code{\link[=moprobit_iter]{moprobit_iter()}}, \code{\link[=chain_to_mcmc]{chain_to_mcmc()}}
}
