% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{moprobit_iter}
\alias{moprobit_iter}
\title{Multivariate ordinal probit MCMC iteration}
\usage{
moprobit_iter(state, iters = 1, TMN_iters = 5, fixSigma = FALSE,
  fixCrossBlockCov = FALSE)
}
\arguments{
\item{state}{A moprobit state container, generated by \code{\link[=moprobit_init]{moprobit_init()}}.}

\item{iters}{The number of iterations to run.}

\item{TMN_iters}{The number of sub-iterations for sampling from the Truncated Multivariate Normal distribution.}

\item{fixSigma}{If \code{TRUE}, the current value of the covariance matrix is taken as known and will not be updated.}

\item{fixCrossBlockCov}{If \code{TRUE}, the covariance for outcomes in separate blocks will be taken as known and will not be updated.}
}
\value{
A new moprobit state container with the current parameters after the requested number of iterations
}
\description{
Run forward n iterations of the Gibbs algorithm starting from \code{state}
}
\seealso{
\code{\link[=moprobit_init]{moprobit_init()}}, \code{\link[=moprobit_chain]{moprobit_chain()}}
}
